Hook before building pages
I am using Grav in combination with another external system that controls the routes which are mapped to Grav pages. As a result of this, I needed to be able to hook the routes BEFORE they are cached. The init() function sets the routes to an empty array, and the routes are cached inside buildPages(). This is why I had to add a new event (onBeforeBuildPages) to enable my own plugin to inject routes externally, also for that I had to adapt the function ->routes() to be able to set a list of routes. I would like the future updates to contain some way of achieving this officially.